[GISel] Move IRTranslatorImpl into cpp file
This is not needed in the header and including it in the header
regresses compile times for PassBuilder.cpp.
